pyproject.toml
[tool.poetry]
name = "accountantiq"
version = "0.1.0"
description = "AccountantIQ monorepo"
authors = ["AccountantIQ <dev@accountantiq.local>"]
readme = "README.md"
packages = [
  { include = "accountantiq_engine", from = "apps/engine" },
  { include = "accountantiq_core", from = "packages/core" },
  { include = "accountantiq_schemas", from = "packages/schemas" },
  { include = "accountantiq_prompts", from = "packages/prompts" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = "^2.6.3"
pyyaml = "^6.0.0"
rapidfuzz = "^3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
types-PyYAML = "^6.0.12"
httpx = "^0.27.0"
mypy = "^1.9.0"
pytest = "^8.1.1"
ruff = "^0.3.7"

[tool.poetry.scripts]
accountantiq-engine = "accountantiq_engine.main:run"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]

[tool.mypy]
python_version = "3.11"
disallow_untyped_defs = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
